pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "discord-py>=2.5.2",
    "gspread>=6.2.0",
    "matplotlib>=3.10.3",
    "oauth2client>=4.1.3",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
